+++ src/lib.rs
@@ -25,13 +25,41 @@
 //!
 //! # Cargo defined Features
 //!
-//! - `default`: Enables the default features of the parser. All of the following features are enabled by default:
-//! - `namespace_removal`: Enables removal of XML namespaces from tag names during parsing.
-//! - `parse_escapes`: Enables parsing of character escapes sequences (`&..;`) in text nodes.
+//! - `default`: Enables the default features of the parser.
+//! - `namespace_removal`: Enables removal of XML namespaces from tag names during parsing. Default is **enabled**.
+//! - `parse_escapes`: Enables parsing of character escapes sequences (`&..;`) in PCData nodes and attribute values. Default is **enabled**.
+//! - `keep_ws_only_pcdata`: all PCData nodes that are composed of whitespace only will be kept. Default is **disabled**.
+//! - `trim_pcdata`: trim whitespaces at beginning and end of PCData nodes. Default is **disabled**.
 //!
-//! # Changelog
+//! # ChangeLog
 //!
-//! See the `CHANGELOG.md` file for further information.
+//! ## [0.2.0] - TBC
+//!
+//! - Going to version [0.2.0] is required as the way that space characters present at the beginning and end of PCData nodes are processed is different, whether or not the following added features are enabled or disabled.
+//! - Added `keep_ws_only_pcdata`: all PCData nodes that are composed of whitespace only will be kept. Default is **disabled**.
+//! - Added `trim_pcdata`: trim whitespaces at beginning and end of PCData nodes. Default is **disabled**.
+//! - Corrected the description of the `parse_escapes` feature to add `attribute values` that are parsed for escapes sequences when that feature is enabled.
+//! - Added test case for each feature. Requires to adjust selected feature before launching the individual tests.
+//!
+//! ## [0.1.2] - 2025-06-12
+//!
+//! - The Document `parser` method is no longer public outside of this crate.
+//! - Added `Nodes` iterator to access document nodes in the sequence of creation. Accessible through the `Document::all_nodes()`, `Document::descendants()` and `Node::descendant()` methods.
+//! - Added blank lines in the doc examples for better readability.
+//! - Adjusted all examples to diminish the required `use` declarations.
+//!
+//! ## [0.1.1] - 2025-06-11
+//!
+//! - Added `pub fn is(&self, name: &str) -> bool` method to `Attribute` and `Node` modules
+//! - Added  `pub use` entries in `lib.rs` to simplify usage in calling applications. All examples and tests have been modified in accordance with this change.
+//! - Added `Display` trait definition for the `ParseXmlError` enum in the `defs` module.
+//!
+//! - Removed the `position` field of the `node_info` struct as the information is available through the range fields of the `NodeType` enum.
+//!
+//! ## [0.1.0] - 2025-06-10
+//!
+//! Initial release
+//!
 pub mod attribute;
 pub mod defs;
 pub mod document;