+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # ChangeLog
 
+## [0.2.0] - TBC
+
+- Going to version [0.2.0] is required as the way that space characters present at the beginning and end of PCData nodes are processed is different, whether or not the following added features are enabled or disabled.
+- Added `keep_ws_only_pcdata`: all PCData nodes that are composed of whitespace only will be kept. Default is **disabled**.
+- Added `trim_ws_pcdata`: trim whitespaces at beginning and end of PCData nodes. Default is **disabled**.
+- Corrected the description of the `parse_escapes` feature to add `attribute values` that are parsed for escapes sequences when that feature is enabled.
+- Added test case for each feature. Requires to adjust selected feature before launching the individual tests.
+
 ## [0.1.2] - 2025-06-12
 
 - The Document `parser` method is no longer public outside of this crate.